+++ R/migrate.R
@@ -13,7 +13,9 @@ migrate <- function() {
     renv_lockfile <- file.path(dir, "renv.lock")
     packrat_lockfile <- file.path(dir, "packrat.lock")
 
-    if (file.exists(renv_lockfile)) {
+    if (is.null(dir)) {
+      message("This project has not yet been packified.")
+    } else if (file.exists(renv_lockfile)) {
       message("renv.lock already exists. You should be good to go.")
     } else if (!file.exists(packrat_lockfile)) {
       message("packrat.lock does not exist.")